+++ src/components/Welcome/Welcome.tsx
@@ -1,7 +1,8 @@
 import classNames from 'classnames/bind';
+
 import styles from './Welcome.module.scss';
 import { Link } from 'react-router-dom';
-import { Button } from '../../ui/Button';
+import { Button } from '../ui/Button';
 
 const cn = classNames.bind(styles);
 
@@ -10,7 +11,7 @@ const Welcome = () => {
         <h1>Transforming your body shape with Power Pulse</h1>
         <Link to='/signin' >
         <Button
-        label="test"
+        label='test'
         action={() => { console.log('click') }}
         customContainerStyles={{marginRight: '12px'}}/>
         </Link>
