+++ src/components/Welcome/index.ts
