--- src/components/config/routes.ts
@@ -1,12 +1,0 @@
-const routes = {
-  HOME_ROUTE: '/',
-  WELCOME_ROUTE: '/welcome',
-  SIGNUP_ROUTE: '/signup',
-  SIGNIN_ROUTE: '/signin',
-  PROFILE_ROUTE: '/profile',
-  DIARY_ROUTE: '/diary',
-  PRODUCTS_ROUTE: '/products',
-  EXERCISES_ROUTE: '/exercises',
-};
-
-export default routes;
