+++ src/components/pages/NotFound/NotFound.tsx
@@ -1,6 +1,7 @@
-import { Link } from 'react-router-dom';
 import classNames from 'classnames/bind';
+
 import styles from './NotFound.module.scss';
+import { Link } from 'react-router-dom';
 
 const cn = classNames.bind(styles);
