+++ src/types/loader.types.ts
@@ -1,0 +1,11 @@
+import { CSSProperties } from 'react';
+
+export type LoaderProps = {
+// visible={true}
+//                 height='120'
+//                 width='120'
+//                 ariaLabel='vortex-loading'
+//                 wrapperStyle={{}}
+//                 wrapperClass='vortex-wrapper'
+//                 colors={['#FC72F7', '#613C67', '#FF73FA', '#73757C', '#2C2C2F', '#181819']}
+};
